requirements.txt
anthropic>=0.34.0
tweepy>=4.14.0
python-dotenv>=1.0.0
pymongo>=4.6.0

# Optional LLM providers (install only what you need)
openai>=1.0.0  # For OpenAI GPT models
google-generativeai>=0.3.0  # For Google Gemini models

